+++ lib/angelo/responder.rb
@@ -1,7 +1,7 @@
 module Angelo
 
   class Responder
-    include Celluloid::Logger
+    include Celluloid::Internals::Logger
 
     class << self
 
@@ -30,8 +30,8 @@ module Angelo
     attr_accessor :connection, :mustermann, :request
     attr_writer :base
 
-    def initialize &block
-      @response_handler = block
+    def initialize method, &block
+      @method, @response_handler = method, block
     end
 
     def reset!
@@ -70,7 +70,7 @@ module Angelo
 
     def handle_error _error, type = :internal_server_error, report = @base.report_errors?
       err_msg = error_message _error
-      Angelo.log @connection, @request, nil, type, err_msg.size
+      Angelo.log @method, @connection, @request, nil, type, err_msg.size
       @connection.respond type, headers, err_msg
       @connection.close
       if report
@@ -170,11 +170,11 @@ module Angelo
         end
       end
 
-      status ||= @redirect.nil? ? :ok : :moved_permanently
-      headers LOCATION_HEADER_KEY => @redirect if @redirect
+      status ||= @redirect.nil? ? :ok : @redirect[1]
+      headers LOCATION_HEADER_KEY => @redirect[0] if @redirect
 
       if @chunked
-        Angelo.log @connection, @request, nil, status
+        Angelo.log @method, @connection, @request, nil, status
         @request.respond status, headers
         err = nil
         begin
@@ -190,7 +190,7 @@ module Angelo
         end
       else
         size = @body.nil? ? 0 : @body.size
-        Angelo.log @connection, @request, nil, status, size
+        Angelo.log @method, @connection, @request, nil, status, size
         @request.respond status, headers, @body
       end
 
@@ -198,9 +198,13 @@ module Angelo
       handle_error e, :internal_server_error
     end
 
-    def redirect url
-      @redirect = url
+    def redirect url, permanent = false
+      @redirect = [url, permanent ? :moved_permanently : :found]
       nil
+    end
+
+    def redirect! url
+      redirect url, true
     end
 
     def on_close= on_close